# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BINAKA GAME API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  password: string;
  country: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface User {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  country: string;
  balance: number;
  vipLevel: number;
  isVerified: boolean;
  isBlocked: boolean;
  isAdmin?: boolean;
  referralCode: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  fullName?: string;
  email?: string;
  country?: string;
  avatarUrl?: string;
}

export interface WalletSummary {
  balance: number;
  totalDeposited: number;
  totalWithdrawn: number;
  totalWon: number;
  totalBet: number;
  pendingDeposits?: number;
  pendingWithdrawals?: number;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  amount: number;
  status: string;
  /** @nullable */
  method: string | null;
  /** @nullable */
  reference?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
}

export interface DepositInput {
  amount: number;
  method: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  walletAddress?: string | null;
}

export interface WithdrawInput {
  amount: number;
  method: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  walletAddress?: string | null;
}

export interface GamePlayInput {
  betAmount: number;
}

export interface ScratchInput {
  betAmount: number;
  cardType: string;
}

export interface SlotsResult {
  reels: string[];
  won: boolean;
  winAmount: number;
  multiplier: number;
  newBalance: number;
  gameId?: number;
}

export interface WheelResult {
  segment: number;
  prize: string;
  won: boolean;
  winAmount: number;
  newBalance: number;
  spinAngle: number;
  gameId?: number;
}

export interface ScratchResult {
  symbols: string[];
  won: boolean;
  winAmount: number;
  newBalance: number;
  cardType: string;
  gameId?: number;
}

export interface GameHistory {
  id: number;
  gameType: string;
  betAmount: number;
  winAmount: number;
  won: boolean;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface GameHistoryList {
  games: GameHistory[];
  total: number;
}

export interface Winner {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  gameType: string;
  winAmount: number;
  /** @nullable */
  multiplier?: number | null;
  createdAt: string;
}

export type WinnerList = Winner[];

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalInvited: number;
  totalEarned: number;
  commissionRate: number;
}

export interface Commission {
  id: number;
  referredUser: string;
  amount: number;
  createdAt: string;
}

export type CommissionList = Commission[];

export interface Notification {
  id: number;
  title: string;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationList {
  notifications: Notification[];
  unreadCount: number;
}

export interface BonusResult {
  amount: number;
  newBalance: number;
  message: string;
}

export interface Bonus {
  id: number;
  title: string;
  description: string;
  amount: number;
  type: string;
  isActive: boolean;
}

export type BonusList = Bonus[];

export interface VipLevel {
  level: number;
  name: string;
  dailyBonus: number;
  cashbackRate: number;
  maxWithdrawal: number;
  badge: string;
  minXp?: number;
}

export interface VipInfo {
  currentLevel: number;
  currentXp: number;
  nextLevelXp: number;
  levels: VipLevel[];
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  totalDeposited: number;
  totalWithdrawn: number;
  platformRevenue: number;
  todaySignups: number;
  todayDeposits: number;
  todayWithdrawals: number;
  totalGamesPlayed: number;
}

export interface AdminUser {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  country: string;
  balance: number;
  vipLevel: number;
  isVerified: boolean;
  isBlocked: boolean;
  isAdmin: boolean;
  referralCode: string;
  createdAt: string;
}

export interface AdminUserList {
  users: AdminUser[];
  total: number;
}

export interface BalanceUpdate {
  amount: number;
  operation: string;
  /** @nullable */
  note?: string | null;
}

export interface StatusUpdate {
  isBlocked: boolean;
}

export interface GameConfig {
  gameType: string;
  minBet: number;
  maxBet: number;
  rtp: number;
  isActive: boolean;
  winProbability?: number;
}

export interface GamesConfig {
  slots: GameConfig;
  wheel: GameConfig;
  scratch: GameConfig;
}

export interface GamesConfigUpdate {
  slots?: GameConfig;
  wheel?: GameConfig;
  scratch?: GameConfig;
}

export type GetTransactionsParams = {
type?: string;
limit?: number;
offset?: number;
};

export type GetSlotsHistoryParams = {
limit?: number;
};

export type GetWheelHistoryParams = {
limit?: number;
};

export type GetScratchHistoryParams = {
limit?: number;
};

export type GetRecentWinnersParams = {
limit?: number;
};

export type GetNotificationsParams = {
limit?: number;
};

export type GetAdminUsersParams = {
search?: string;
limit?: number;
offset?: number;
};

export type GetAdminTransactionsParams = {
status?: string;
type?: string;
limit?: number;
offset?: number;
};