# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BINAKA GAME API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}/balance:
    patch:
      operationId: updateUserBalance
      tags: [admin]
      summary: Update user balance (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Block/unblock user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/games/config:
    get:
      operationId: getGamesConfig
      tags: [admin]
      summary: Get games configuration (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamesConfig"
    patch:
      operationId: updateGamesConfig
      tags: [admin]
      summary: Update games configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GamesConfigUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamesConfig"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: Get all transactions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /admin/transactions/{id}/approve:
    patch:
      operationId: approveTransaction
      tags: [admin]
      summary: Approve a transaction (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /admin/transactions/{id}/reject:
    patch:
      operationId: rejectTransaction
      tags: [admin]
      summary: Reject a transaction (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [fullName, phone, password, country]
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        password:
          type: string
        country:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, fullName, phone, country, balance, vipLevel, isVerified, isBlocked, referralCode, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        country:
          type: string
        balance:
          type: number
        vipLevel:
          type: integer
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        isAdmin:
          type: boolean
        referralCode:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Transaction:
      type: object
      required: [id, userId, type, amount, status, method, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        method:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionList:
      type: object
      required: [transactions, total]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
...
    AdminUser:
      type: object
      required: [id, fullName, phone, country, balance, vipLevel, isVerified, isBlocked, isAdmin, referralCode, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        country:
          type: string
        balance:
          type: number
        vipLevel:
          type: integer
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        isAdmin:
          type: boolean
        referralCode:
          type: string
        createdAt:
          type: string
    AdminUserList:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
    BalanceUpdate:
      type: object
      required: [amount, operation]
      properties:
        amount:
          type: number
        operation:
          type: string
        note:
          type: ["string", "null"]
    StatusUpdate:
      type: object
      required: [isBlocked]
      properties:
        isBlocked:
          type: boolean
    GameConfig:
      type: object
      required: [gameType, minBet, maxBet, rtp, isActive]
      properties:
        gameType:
          type: string
        minBet:
          type: number
        maxBet:
          type: number
        rtp:
          type: number
        isActive:
          type: boolean
        winProbability:
          type: number
    GamesConfig:
      type: object
      required: [slots, wheel, scratch]
      properties:
        slots:
          $ref: "#/components/schemas/GameConfig"
        wheel:
          $ref: "#/components/schemas/GameConfig"
        scratch:
          $ref: "#/components/schemas/GameConfig"
    GamesConfigUpdate:
      type: object
      properties:
        slots:
          $ref: "#/components/schemas/GameConfig"
        wheel:
          $ref: "#/components/schemas/GameConfig"
        scratch:
          $ref: "#/components/schemas/GameConfig"