# lib/api-zod/src/generated/types/gamesConfigUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BINAKA GAME API
 * OpenAPI spec version: 0.1.0
 */
import type { GameConfig } from './gameConfig';

export interface GamesConfigUpdate {
  slots?: GameConfig;
  wheel?: GameConfig;
  scratch?: GameConfig;
}