# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BINAKA GAME API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  country: string;
  balance: number;
  vipLevel: number;
  isVerified: boolean;
  isBlocked: boolean;
  isAdmin?: boolean;
  referralCode: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}